services/pyproject.toml
[project]
name = "rhizomatic-services"
version = "0.0.0"
description = "Python processing services for the Rhizomatic knowledge base"
requires-python = ">=3.12"
dependencies = [
    "redis>=5.0.0",
    "pydantic>=2.9.0",
]

[project.optional-dependencies]
processor = [
    "spacy>=3.8.0",
    "trafilatura>=1.12.0",
]
embedder = [
    "sentence-transformers>=3.3.0",
    "torch>=2.4.0",
]
ocr = [
    "pytesseract>=0.3.13",
    "Pillow>=10.4.0",
]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
